concept_2/eplf/requirements.txt
pika==1.2.0
psycopg2-binary==2.9.1
schwifty==2023.6.0
